+++ CHANGELOG.md
@@ -7,6 +7,11 @@ Performance improvements:
 * [PERF #1381](https://github.com/BurntSushi/ripgrep/pull/1381):
   Directory traversal is sped up with speculative ignore-file existence checks.
 
+Feature enhancements:
+
+* [FEATURE #1390](https://github.com/BurntSushi/ripgrep/pull/1390):
+  Add new `--no-context-separator` flag that always hides context separators.
+
 Bug fixes:
 
 * [BUG #1335](https://github.com/BurntSushi/ripgrep/issues/1335):